move Spinner
