+++ webui/react/src/components/Markdown.tsx
@@ -3,7 +3,7 @@ import { default as MarkdownViewer } from 'markdown-to-jsx';
 import React, { useCallback } from 'react';
 
 import css from './Markdown.module.scss';
-import Spinner from './Spinner';
+import Spinner from '../tmp-shared/components/Spinner/Spinner';
 
 const { TabPane } = Tabs;
 const MonacoEditor = React.lazy(() => import('components/MonacoEditor'));
