+++ webui/react/src/hooks/useModal/useJupyterLabModal.tsx
@@ -5,7 +5,7 @@ import yaml from 'js-yaml';
 import React, { Dispatch, useCallback, useEffect, useMemo, useReducer, useState } from 'react';
 
 import Link from 'components/Link';
-import Spinner from 'components/Spinner';
+import Spinner from 'tmp-shared/components/Spinner/Spinner';
 import useJupyterLab from 'hooks/useJupyterLab';
 import usePrevious from 'hooks/usePrevious';
 import useStorage from 'hooks/useStorage';
