+++ webui/react/src/tmp-shared/components/Spinner/Spinner.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import HelmetDecorator from 'storybook/HelmetDecorator';
 import StoreDecorator from 'storybook/StoreDecorator';
 
-import Page from './Page';
+import Page from '../../../components/Page';
 import Spinner from './Spinner';
 
 export default {
