+++ webui/react/src/tmp-shared/components/Spinner/index.ts
@@ -1,0 +1,3 @@
+export * from './Spinner';
+import Spinner from './Spinner';
+export default Spinner;
